scaler

+++ train.py
@@ -148,7 +148,7 @@ def train(
 
             # batch = {k: v.to(device) for k, v in batch.items()}
 
-            x = batch['input_ids'].to(device, dtype=torch.float)
+            x = batch['input_ids'].to(device)
             with torch.cuda.amp.autocast(enabled=fp16):
                 loss = model(x, return_loss=True)
                 std = 0
